Add Clang compatibility and fix remaining Clang-specific warnings
- Add Clang-specific suppression for -Wvariadic-macro-arguments-omitted
  (triggered by CUDA toolkit headers) and -Wsign-conversion (not part
  of GCC's -Wconversion)
- Fix deprecated literal operator syntax (remove space before _GiB etc.)
- Fix Clang -Wdouble-promotion: use lroundf() instead of lround() for
  float args, explicit casts for long double expressions
- Fix Clang -Wfloat-conversion: use explicit != 0.0F instead of
  implicit float-to-bool
- Fix Clang -Wsign-conversion: add static_cast<size_t>() for int-to-
  size_t conversions in sample code

+++ plugin/groupNormalizationPlugin/groupNormalizationPlugin.cpp
@@ -159,7 +159,7 @@ int32_t GroupNormalizationPlugin::enqueue(nvinfer1::PluginTensorDesc const* inpu
         mBNTensorDesc,                 //
         mBnScales->mPtr,               // 1
         mBnBias->mPtr,                 // 0
-        0.0F,                          // exponential average factor
+        static_cast<double>(0.0F),     // exponential average factor
         nullptr,                       // resultRunningMean
         nullptr,                       // resultRunningVar
         static_cast<double>(mEpsilon), //  eps
